remove extra test suite
