+++ packages/kit/test/apps/options-2/package.json
@@ -10,7 +10,7 @@
 		"check": "svelte-kit sync && tsc && svelte-check",
 		"test": "pnpm test:dev && pnpm test:build",
 		"test:dev": "DEV=true playwright test",
-		"test:build": "NODE_ENV=custom playwright test && REGISTER_SERVICE_WORKER=true playwright test"
+		"test:build": "NODE_ENV=custom DYNAMIC_PUBLIC_ENV=true playwright test && REGISTER_SERVICE_WORKER=true playwright test"
 	},
 	"devDependencies": {
 		"@sveltejs/adapter-node": "workspace:^",
