+++ packages/kit/test/apps/options-2/test/test.js
@@ -1,49 +1,10 @@
-import fs from 'node:fs';
 import process from 'node:process';
-import { fileURLToPath } from 'node:url';
 import { expect } from '@playwright/test';
 import { test } from '../../../utils.js';
-
-const output = fileURLToPath(new URL('../.svelte-kit/output', import.meta.url));
 
 test.skip(() => !!process.env.REGISTER_SERVICE_WORKER);
 
 test.describe.configure({ mode: 'parallel' });
-
-test.describe('env', () => {
-	test('resolves upwards', async ({ page }) => {
-		await page.goto('/basepath/env');
-		expect(await page.textContent('[data-testid="public"]')).toBe('public: hello');
-		expect(await page.textContent('[data-testid="private-dynamic"]')).toBe(
-			'private dynamic: secret resolved at runtime'
-		);
-		expect(await page.textContent('[data-testid="private-static"]')).toBe(
-			'private static: secret resolved at build time'
-		);
-		expect(await page.textContent('[data-testid="private-validated-default"]')).toBe(
-			'private validated default: foo'
-		);
-	});
-
-	test('applies explicit env vars to %sveltekit.env%', async ({ page }) => {
-		await page.goto('/basepath');
-		await expect(page.locator('body')).toHaveAttribute('data-message', 'hello');
-	});
-
-	test('correct values are exported from $app/env/*', async ({ page }) => {
-		await page.goto('/basepath/env/import-all');
-
-		await expect(page.locator('[data-private]')).toHaveText(
-			JSON.stringify({
-				PRIVATE_EXPLICIT_ENV: 'secret resolved at runtime',
-				PRIVATE_STATIC_EXPLICIT_ENV: 'secret resolved at build time',
-				PRIVATE_VALIDATED_DEFAULT_ENV: 'foo'
-			})
-		);
-
-		await expect(page.locator('[data-public]')).toHaveText(JSON.stringify({ MESSAGE: 'hello' }));
-	});
-});
 
 test.describe('paths', () => {
 	test('serves /basepath', async ({ page }) => {
@@ -186,37 +147,6 @@ test.describe("bundleStrategy: 'single'", () => {
 	test('serialization works with streaming', async ({ page }) => {
 		await page.goto('/basepath/serialization-stream');
 		await expect(page.locator('h1', { hasText: 'It works!' })).toBeVisible();
-	});
-});
-
-test.describe('$app/env', () => {
-	// regression test for https://github.com/sveltejs/kit/issues/15971:
-	// importing `$app/env` before the router is initialized (e.g. in
-	// hooks.client.js) must not throw `__SVELTEKIT_APP_VERSION__ is not defined`
-	test('version is defined when imported during client init', async ({
-		page,
-		javaScriptEnabled
-	}) => {
-		test.skip(!javaScriptEnabled || !process.env.DEV);
-
-		await page.goto('/basepath', { wait_for_started: false });
-		await expect(page.locator('body.started')).toBeVisible();
-		expect(await page.pageErrors()).toHaveLength(0);
-	});
-
-	test('loads dynamic public environment variables in the service worker', ({
-		javaScriptEnabled
-	}) => {
-		test.skip(javaScriptEnabled || !process.env.DEV);
-
-		const content = fs.readFileSync(
-			`${output}/prerendered/dependencies/_app/env.script.js`,
-			'utf-8'
-		);
-		expect(content).toContain('hello');
-
-		const service_worker = fs.readFileSync(`${output}/client/service-worker.js`, 'utf-8');
-		expect(service_worker).toContain('importScripts("/basepath/_app/env.script.js");');
 	});
 });
