--- packages/kit/test/prerendering/explicit-env-static/.env
@@ -1,1 +1,0 @@
-STATIC_MESSAGE=the static message
