--- packages/kit/test/prerendering/explicit-env-static/.gitignore
@@ -1,1 +1,0 @@
-!.env
