--- packages/kit/test/prerendering/explicit-env-static/package.json
@@ -1,23 +1,0 @@
-{
-	"name": "prerendering-test-explicit-env-static",
-	"private": true,
-	"version": "0.0.1",
-	"scripts": {
-		"dev": "vite dev",
-		"build": "vite build",
-		"preview": "vite preview",
-		"prepare": "svelte-kit sync",
-		"check": "svelte-kit sync && tsc && svelte-check",
-		"test": "svelte-kit sync && pnpm build && vitest run"
-	},
-	"devDependencies": {
-		"@sveltejs/kit": "workspace:^",
-		"@sveltejs/vite-plugin-svelte": "catalog:",
-		"svelte": "catalog:",
-		"svelte-check": "catalog:",
-		"typescript": "catalog:",
-		"vite": "catalog:",
-		"vitest": "catalog:"
-	},
-	"type": "module"
-}
