--- packages/kit/test/prerendering/explicit-env-static/src/app.html
@@ -1,11 +1,0 @@
-<!doctype html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1" />
-		%sveltekit.head%
-	</head>
-	<body>
-		%sveltekit.body%
-	</body>
-</html>
