--- packages/kit/test/prerendering/explicit-env-static/src/env.js
@@ -1,6 +1,0 @@
-export const variables = {
-	STATIC_MESSAGE: {
-		public: true,
-		static: true
-	}
-};
