--- packages/kit/test/prerendering/explicit-env-static/src/routes/+page.svelte
@@ -1,5 +1,0 @@
-<script>
-	import { STATIC_MESSAGE } from '$app/env/public';
-</script>
-
-<h1>{STATIC_MESSAGE}</h1>
