--- packages/kit/test/prerendering/explicit-env-static/src/service-worker.js
@@ -1,3 +1,0 @@
-import { STATIC_MESSAGE } from '$app/env/public';
-
-console.log(STATIC_MESSAGE);
