--- packages/kit/test/prerendering/explicit-env-static/svelte.config.js
@@ -1,14 +1,0 @@
-import adapter from '../../../../adapter-static/index.js';
-
-/** @type {import('@sveltejs/kit').Config} */
-const config = {
-	kit: {
-		adapter: adapter(),
-
-		experimental: {
-			explicitEnvironmentVariables: true
-		}
-	}
-};
-
-export default config;
