--- packages/kit/test/prerendering/explicit-env-static/vite.config.js
@@ -1,23 +1,0 @@
-import * as path from 'node:path';
-import { sveltekit } from '@sveltejs/kit/vite';
-
-/** @type {import('vite').UserConfig} */
-const config = {
-	build: {
-		minify: false
-	},
-
-	clearScreen: false,
-
-	logLevel: 'silent',
-
-	plugins: [sveltekit()],
-
-	server: {
-		fs: {
-			allow: [path.resolve('../../../src')]
-		}
-	}
-};
-
-export default config;
